fix(component) : refactor component gits button
