--- lib/pages/component/gits_button/src/icon_position.dart
@@ -1,1 +1,0 @@
-enum IconPosition { start, end }
